modules/watsonx-data/src/test/java/com/ibm/cloud/watsonxdata/watsonx_data/v2/model/AddSparkEngineCatalogsOptionsTest.java
/*
 * (C) Copyright IBM Corp. 2024.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations under the License.
 */

package com.ibm.cloud.watsonxdata.watsonx_data.v2.model;

import com.ibm.cloud.sdk.core.service.model.FileWithMetadata;
import com.ibm.cloud.watsonxdata.watsonx_data.v2.model.AddSparkEngineCatalogsOptions;
import com.ibm.cloud.watsonxdata.watsonx_data.v2.utils.TestUtilities;
import java.io.InputStream;
import java.util.HashMap;
import java.util.List;
import org.testng.annotations.Test;
import static org.testng.Assert.*;

/**
 * Unit test class for the AddSparkEngineCatalogsOptions model.
 */
public class AddSparkEngineCatalogsOptionsTest {
  final HashMap<String, InputStream> mockStreamMap = TestUtilities.createMockStreamMap();
  final List<FileWithMetadata> mockListFileWithMetadata = TestUtilities.creatMockListFileWithMetadata();

  @Test
  public void testAddSparkEngineCatalogsOptions() throws Throwable {
    AddSparkEngineCatalogsOptions addSparkEngineCatalogsOptionsModel = new AddSparkEngineCatalogsOptions.Builder()
      .engineId("testString")
      .catalogNames("testString")
      .authInstanceId("testString")
      .build();
    assertEquals(addSparkEngineCatalogsOptionsModel.engineId(), "testString");
    assertEquals(addSparkEngineCatalogsOptionsModel.catalogNames(), "testString");
    assertEquals(addSparkEngineCatalogsOptionsModel.authInstanceId(), "testString");
  }

  @Test(expectedExceptions = IllegalArgumentException.class)
  public void testAddSparkEngineCatalogsOptionsError() throws Throwable {
    new AddSparkEngineCatalogsOptions.Builder().build();
  }

}